Move agent integration test helpers into their own file
